Fix All Bugs

+++ database/migrations/2024_12_13_195825_create_services_table.php
@@ -9,7 +9,7 @@ class CreateServicesTable extends Migration
     public function up()
     {
         Schema::create('services', function (Blueprint $table) {
-            $table->integer('id')->primary();
+            $table->unsignedBigInteger('id')->primary();
             $table->string('name');
             $table->string('type');
             $table->foreignId('category_id')->nullable()->constrained('categories')->onDelete('set null');
